Add FreeBSD support
FreeBSD uses inotifywait(1) like Linux.

The only change is the oneliner used to call it: argument parsing is
stopped with a `--` just after `-c "<command line>"` on FreeBSD,
otherwise sh(1) eats `-m` and `-e`. That change is made specific to
FreeBSD. It is possible this would work on Linux too, but I don't have
all the Dash/Bash/etc. implementations at hands, so let's not break
something which works for years.